+++ library/laboratory/src/main/java/io/mehow/laboratory/FeatureFactory.kt
@@ -10,5 +10,13 @@ public interface FeatureFactory {
    */
   public fun create(): Set<Class<Feature<*>>>
 
+  /**
+   * Creates a new [FeatureFactory] that will return a combined set of this factory and the other factory.
+   */
+  @JvmDefault
+  public operator fun plus(factory: FeatureFactory): FeatureFactory = object : FeatureFactory {
+    override fun create() = this@FeatureFactory.create() + factory.create()
+  }
+
   public companion object
 }