+++ library/docs/changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `FeatureFactory` can be now appended to another with plus operator.
+
 ### Changed
 - Use description as feature flag's KDoc content.
 - Upgrade Kotlin to `1.4.21`.